BUG: Refs #209.  Fixed a failing test, and made other improvements from Charles.
There was a failing test because I was referring to "global" from the ini file,
where the "global" key didn't exist.  Charles gave me many good suggestions
in QAing this original work, some of which I have added in.  Also, I've moved
some functionality from the ConfigController to the KWBatchmakeComponent in
an attempt to allow them to be reused from other places in the module.