BUG: Refs #212.  Rewrite of KWBatchmakeComponent.php, and reworking of tests.
Rewrote KWBatchmakeComponent.php to be more sensible, rewrote tests and other
code that depending on this.  Also added a README to the batchmake module for
testing, and a switch in CMakeLists.txt for the batchmake module to have
batchmake module testing be off by default, since a certain amount of config
is required to have the batchmake tests pass.