ENH: refs #0257. Adding a notification class that includes web-api hooks.
This base class will enable users to define web-api methods in an API
component and then easily include them in the web-api through their
module notification class without having to worry about a lot of
callback and setup. In fact, the api component of their module will be
parsed automatically.