ENH: refs #0258. changed loadData function to allow filename and model modules.
Before this fix, it was not possible to have a test xml dataset in modules that
had core models.  This fix allows calling loadData with both a fileModule, which
can specify the module that the data file came from, and a modelModule param,
which can specify the module that the model came from.  E.g.
  loadData('User','default','','batchmake');
would load a file default.xml defined in the batchmake module that has
User models from the core defined in it, with this command loading those users.
  loadData('Itemmetric','default','batchmake','batchmake');
would load a file default.xml defined in the batchmake module that has
Itemmetric models defined in the batchmake module, with this command loading
those itemmetrics.