ENH: refs #0236. Forbid duplicate content uploads via web API
midas.upload.generatetoken now requires a checksum parameter.  If we already
have a bitstream with the provided checksum, we don't create an upload token;
instead we create a pointer to the existing content within the specified item
and return an empty token to let the user know that they don't need to call
the upload.process function.