BUG: refs #0283. Got this setup working for postgres.
I had not tested the setup for postgres, and the previous setup
actually copied a mysql.ini testing config file over the pgsql.ini
testing config file.  This version fixes that, and some other bugs
that I found when cleaning up the pgsql test and getting them to work.
Furthermore, there are some fixes and improvements to the helloworld module
that were useful in testing this script.