BUG: Refs #0283. Fixed AssetstoreModelTest failures by adding type juggling.
Pgsql and Mysql have different behaviors related to the type of data, so
in order to use compareDao in these tests across the two db systems I added
an option to MIDASModel to compare dao objects with typejuggling, leaving
the default behavior to not type juggle.  These tests call compareDao with
typejuggling.