BUG: refs #0326. called setupDatabase on default dataset for pgsql.
If you load rows from the default.xml dataset, and then try to programmatically
create an object of the same type, in postgres an exception will be thrown
because the mechanism for loading from default.xml does not update the sequence
values in postgres (because the default.xml already supplies ids), so when
you next try to create an object of the same type, postgres gives you the next
sequence value as your id, which is in conflict with a row in the db.  The way
around this is to update the sequence value with the current max id, and this
will occur when you call setupDatabase for the dataset in question.