BUG: refs #0326. Add an explicit case to deleting a linked feed, for pgsql.
Postgres will not automatically do the type conversion from integer to string,
which causes a problem when we delete an item and we are trying to find feed
rows that have as their ressource the deleted item's id.  So for pgsql we
must explicitly cast the itemId to a string, to match the varchar type
of the ressource column in the feed table.