BUG: refs #0349. Item create description is optional, fixed null revision ref.
There was an additional instance of the revision variable being set to null
that was fixed, and now the item description being not included won't break the
item create api call.  Extensive testing of all fixes in this topic branch
are included in the changes to the ApiCallMethodsTest, in fact, the extensive
testing is what led to many of these small bugs being found.