BUG: refs #0373. Cleaned signature of createNewRevision method.
Cleaned up the signature of the createNewItemRevision method, and made some
changes in related areas for improved readability.  I found that the number
of parameters in the new signature violated a style check, which is probably
why the signature is as it was, but that seems like an arbitrary limit, and
I don't think it should get in the way of clarity of a method signature.