BUG: refs #0381. Cleaned out error supression.
Since we have error reporting turned on for STRICT, and we are setting
fields on a field that itself has not yet been created, we were generating
error messages.  By explicitly creating the field if the field does not yet
exist, the error message isn't generated, thus error supression can be removed.