BUG: refs #0380. Changed getTempDirectory to UtilityComponent.
Took the getTempDirectory out of KWUtils, put it into the UtilityCompnent,
and changed all references to the getTempDirectory, or string literals
pointing to BASE_PATH . /tmp to this function.  Also added a convenience
function call in the GlobalController and MIDASModel so that all subclasses
would have easy access to the functionality.