ENH: refs #0277. Removed the code that handled $_GET['reset'].
This was attempting to drop the database and recreate it. This would
only work if the person installing Midas was using a user that can
create databases as the midas user (or their root user doesn't have a
password set). Because both of these situations are discouraged in any
deployment of the system, I've removed the offending code.