BUG: refs #453. Admins can now see all private feeds
Also fixed a bug where feed related to a community invitation was not removed
when the community invitation was deleted as a result of user deletion, which
caused the feed page to crash. Admins will see the recipient of a community
invitation in the feed now instead of just seeing "invited you to join the
community".