BUG: refs #0530. Saved item upon item change, removed stray metadata delete.
When an item's metadata changes, when a metadatavalue is added, removed, or
edited, the Item will then be saved to update the Lucene index.

Also, there was a stray bit of code to delete metadata that was never being
executed, this has been cleaned out.  Deleting metadata through the correct
mechanism still works.