ENH: refs #0586. Style fix, added test, and added method removeRevision.
It makes more sense to have the code to remove a revision in the Item model,
so this was created and called from the ItemController.  This method was also
tested in the ItemModelTest.  Rather than having the ItemRevisionModel
update the size of the Item when a revision is removed from the Item, I had
the Item Model take care of that in the newly introduced removeRevision method.