ENH: refs #0588. Enhancing community descriptions for style and consistency.
To do this I added support for Markdown in the descriptions and added
JQuery dotdotdot to shorten the text with javascript and link the user
to the full, rendered description. One trick for the shortened
description is rendering the markdown and then stripping tags. This
ends up capturing only the text (which is desired behavior) for the
description snapshots on the folder view and comunity list.