ENH: refs #727. Allow modules to intercept successful auth events
Before the auth is committed to the session.  Mfa uses this to write a
temp user record into a different session namespace and pop open a dialog
prompting the user for his OTP, which is then authenticated. On success,
the temp session entry is converted into the "logged in" expected entry.