ENH: refs #416. Remove version comparison in admincontroller
We had added this because it was impossible to upgrade via the UI when
moving from version 3.1.3 to 3.1.4 because some of the code in the admin
controller references code that requires version 3.1.4 to work.  There are
no instances still running versions prior to 3.1.4, so it is safe to
remove this check now.