ENH: refs #0847. Added tests that exposed privacy bugs in item.create.
There was a bug when changing privacy in an already created item and a bug
where when privacy 'Public' was passed to a item to be created it was
not getting set correctly.  Plus the default privacy when creating a new item
was not getting set correctly. The additions to this test exposes these bugs.