ENH: refs #0847. Fixed bugs in item.create privacy.
Fixed the bugs when creating a new item to set the default privacy correctly,
and to set the privacy correctly when creating a new item and passing in
privacy value.  Also setting the privacy value correctly when modifying an
existing item.  Plus a small change to folder.create that set the default
privacy excplitly in the API method.