ENH: refs #982. CMake script for creating a new skeleton module
This script will create a very minimal skeleton of a module.  The example
template can definitely be fleshed out and improved in the future to
contain more examples of common things people do in modules, e.g. loading
a dao using a model class, implementing a pdo function for searching
records by a certain field, etc.