BUG: refs #1009. Fixing the item rename (for counting).
This is now done without a preg_split and preg_match. We use the
database to check for redundancy. This could degenerate into a long
loop, but this inefficiency is outweighed in both simplicity (compared
to a single LIKE query and counting) and the efficiency gain of not
iterating though all of the item's siblings.