Improve formatting of tracker module scalar details view.
The scalar details view displays rows of metrics with units.  This
commit rounds the float value of the scalars to 4 decimal places, aligns
all of the values to the right for display, and splits out the units to
a separate table column so they can also be right aligned for display.
If the values and units were left in the same column it would be more
difficult to align.